feat: add external-posts, refactor
