+++ external-posts.json
@@ -1,0 +1,7 @@
+[
+  {
+    "externalLink": "https://youtu.be/9yFi3-cfprQ",
+    "date": "2021-10-10",
+    "title": "JSMD Meetup #12: Dragoș Străinu / You Don't Know *.user.js"
+  }
+]
