+++ lib/posts.ts
@@ -6,58 +6,12 @@ import { remark } from "remark";
 import html from "remark-html";
 import prism from "remark-prism";
 import remarkGfm from "remark-gfm";
+import {
+  getMDContent,
+  postsDirectory,
+} from "../scripts/utils";
+import { PostData, PostFolderName } from "./types";
 
-const postsDirectory = path.join(process.cwd(), "posts");
-
-function getMDContent(markdownFile: string): string {
-  // Read markdown file as string
-  const fullPath = path.join(postsDirectory, markdownFile);
-  const fileContents = fs.readFileSync(fullPath, "utf8");
-  return fileContents;
-}
-
-export type PostData = {
-  date: Date | string;
-  title: string;
-  description: string;
-  draft?: boolean;
-  starred?: boolean;
-  banner?: string;
-  slug: string;
-  tags: string[];
-  contentHtml: string;
-};
-
-export function getSortedPostsData(): PostData[] {
-  // Get file names under /posts
-  const fileNames = fs.readdirSync(postsDirectory);
-  const allPostsData: PostData[] = fileNames.map((folderName) => {
-    const fileContents = getMDContent(folderName);
-
-    // Use gray-matter to parse the post metadata section
-    const matterResult = matter(fileContents);
-
-    // Combine the data with the id
-    return {
-      ...matterResult.data,
-      slug: matterResult.data.slug || kebabCase(matterResult.data.title),
-    } as PostData;
-  });
-  // Sort posts by date
-  return allPostsData
-    .filter((p) => !p.draft)
-    .sort(({ date: a }, { date: b }) => {
-      if (a < b) {
-        return 1;
-      } else if (a > b) {
-        return -1;
-      } else {
-        return 0;
-      }
-    });
-}
-
-export type PostFolderName = { folderName: string; postSlug: string };
 
 export function getAllPostFolderNames(): PostFolderName[] {
   // Get file names under /posts
@@ -108,3 +62,5 @@ export async function getPostData(folderName: string): Promise<PostData> {
     contentHtml,
   } as PostData;
 }
+
+export { getSortedPostsData } from "../scripts/utils";
